Fix a few conversion warnings
git-svn-id: svn://svn.code.sf.net/p/synce/code/trunk@4124 599ab11a-0111-0410-8a50-d7a5b0953570

+++ core/lib/rapi/backend_ops_1/invoke.c
@@ -116,7 +116,7 @@ static HRESULT CeRapiInvokeBuffers(
   HRESULT return_value = E_UNEXPECTED;
   HRESULT hr;
   uint32_t unknown;
-  unsigned bytes_left;
+  size_t bytes_left;
   unsigned output_size;
   uint32_t last_error;
